+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Import as GoodNews subscribers if GoodNews add-on is installed
+- Assign GoodNews groups and categories on import
+
 ## [2.0.0-beta1]
 ### Added
 - Added new setting mail_format (HTML or Plain text)